// src/modules/protocols/domain/enums/protocol-status.enum.ts
export enum ProtocolStatus {
  INICIADO = 9,
  COMPLETO = 10,
  INCOMPLETO = 11,
  ARCHIVADO_VENCIMIENTO = 12,
  EN_EVALUACION = 13,
  EVALUADO = 14,
  EN_REVISION_SECRETARIA = 15,
  DISCREPANCIA_RIESGO = 16,
  APROBADO = 17,
  RECHAZADO = 18,
  REQUIERE_SUBSANACION_VERSION = 19,
  REQUIERE_SUBSANACION_DOC = 20,
  EN_CONTROL_DOCUMENTAL = 21,
}